airflow/requirements.txt
dbt-core>=1.6.0
dbt-clickhouse>=1.4.0
clickhouse-connect>=0.6.0
pandas>=2.0.0
openpyxl>=3.1.0
kafka-python>=2.0.2
cryptography>=41.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
kafka-python>=2.0.2